% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{covariance_mats}
\alias{covariance_mats}
\title{Covariance Matrix
gets covariance_matrix of K and Ks}
\usage{
covariance_mats(X, Xt, k)
}
\arguments{
\item{X}{training inputs}

\item{Xt}{testing inputs}

\item{k}{covariance function}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{K covariance matrix of training input
Ks covariance matrix of trainign data and test data
}\if{html}{\out{</div>}}
}
\description{
Covariance Matrix
gets covariance_matrix of K and Ks
}
\examples{
k <- function(r) {
  return(exp(-0.5 * (r / 4)^2))
}
n <- 10
p <- 3
nt <- 4
X <- matrix(rnorm(n * p, 0, 0.3), n, p)
Xt <- matrix(rnorm(nt * p, 0, 0.3), nt, p)

}
