% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{exp_kernel}
\alias{exp_kernel}
\title{Exponential kernel}
\usage{
exp_kernel(l, r = 1)
}
\arguments{
\item{l}{length scale}

\item{r}{euclidean distance between two data points}
}
\value{
covariance function with fixed parameter l
}
\description{
Exponential kernel
}
\examples{
exp_kernel(3)
exp_kernel(3.5)
}
