% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{get_r}
\alias{get_r}
\title{Returns Euclidean distance between two input data}
\usage{
get_r(x1, x2)
}
\arguments{
\item{x1}{first data}

\item{x2}{second data}
}
\value{
r euclidean distance between first data and the second data
}
\description{
Returns Euclidean distance between two input data
}
\examples{
x1 <- matrix(rnorm(9))
x2 <- matrix(rnorm(9))
get_r(x1, x2)
}
