% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gpr_cv}
\alias{gpr_cv}
\title{Gaussian Process Regression Cross Validation}
\usage{
gpr_cv(X, y, k, sigma2, num_folds, fold_ids = NULL)
}
\arguments{
\item{X}{training inputs}

\item{y}{training targets}

\item{k}{list of kernels}

\item{sigma2}{noise level}

\item{num_folds}{number of folds}

\item{fold_ids}{list of ids to indicate folds}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   cvm average mean square errors for each kernels
   k original kernels
}\if{html}{\out{</div>}}
}
\description{
Gaussian Process Regression Cross Validation
}
\examples{
n <- 100
p <- 3
nt <- 4
X <- matrix(rnorm(n * p, 0, 0.3), n, p)
y <- matrix(rnorm(n), n, 1)
sigma2 <- 10
k1 <- function(r) {
  return(exp(-0.5 * (r / 4)^2))
}

k2 <- function(r) {
  return(exp(-r / 3.5))
}
k <- c(k1, k2)
num_folds <- 5
gpr_cv(X, y, k, sigma2, num_folds)
}
