% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gpr_seq_kernels}
\alias{gpr_seq_kernels}
\title{Gaussian Process Regression for a sequence of kernels}
\usage{
gpr_seq_kernels(X, y, k, sigma2, Xt, yt)
}
\arguments{
\item{X}{training inputs}

\item{y}{training targets}

\item{k}{list of covariance functions}

\item{sigma2}{noise level}

\item{Xt}{testing inputs}

\item{yt}{testing targets}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   k: list of covariance functions, same as input
   mse: mean square errors for each covariance function
   predictions: [nt x num_kernels] matrix containing predictions of each test inputs under each kernels
}\if{html}{\out{</div>}}
}
\description{
Gaussian Process Regression for a sequence of kernels
}
\examples{
n <- 10
p <- 3
nt <- 4
X <- matrix(rnorm(n * p, 0, 0.3), n, p)
y <- matrix(rnorm(n), n, 1)
Xt <- matrix(rnorm(nt * p, 0, 0.3), nt, p)
yt <- matrix(rnorm(nt), nt, 1)
num_folds <- 3
sigma2 <- 10


k1 <- function(r) {
  return(exp(-0.5 * (r / 4)^2))
}

k2 <- function(r) {
  return(exp(-r / 3.5))
}
k <- c(k1, k2)
gpr_seq_kernels(X, y, k, sigma2, Xt, yt)

}
