% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{matern_kernel}
\alias{matern_kernel}
\title{Matern kernel}
\usage{
matern_kernel(l, v, r = 1)
}
\arguments{
\item{l}{length-scale}

\item{v}{smoothness}

\item{r}{euclidean distance between two data points}
}
\value{
covariance function with fixed parameter l and v
}
\description{
Matern kernel
}
\examples{
matern_kernel(3, 2.5)
matern_kernel(1, 1.5)
}
