% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{pick_kernel}
\alias{pick_kernel}
\title{pick the kernel
return kernel functions with hyperparemeters}
\usage{
pick_kernel(para_list, method = c("se", "m", "exp"))
}
\arguments{
\item{para_list}{a list of parameter that will be pass on to the kernel function}

\item{method}{a string indicates what function class to pick}
}
\value{
return one of the kernel functions with input parameter list
}
\description{
pick the kernel
return kernel functions with hyperparemeters
}
\examples{
pick_kernel(list(1, 2), "m")
pick_kernel(list(3), "se")
}
