% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{se_kernel}
\alias{se_kernel}
\title{Square Exponential Kernel}
\usage{
se_kernel(l, r = 1)
}
\arguments{
\item{l}{length scale parameter of squared exponential kernel}

\item{r}{euclidean distance between two data points}
}
\value{
covariance function with fixed hyperperameter l
}
\description{
Square Exponential Kernel
}
\examples{
l = 3
se_kernel(3)
}
