% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{standardize}
\alias{standardize}
\title{Standardize Input
Standardize X, y, Xt, and yt based on mean and variance of X and y}
\usage{
standardize(X, y, Xt, yt)
}
\arguments{
\item{X}{original training inputs}

\item{y}{original training targets}

\item{Xt}{original testing inputs}

\item{yt}{original testing targets}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   X scaled training inputs
   y scaled training targets
   Xt scaled testing inputs
   yt scaled testing targets
}\if{html}{\out{</div>}}
}
\description{
Standardize Input
Standardize X, y, Xt, and yt based on mean and variance of X and y
}
\examples{
n = 10
p = 3
nt = 4
X <- matrix(rnorm(n*p, 0, 0.3), n, p)
y = matrix(rnorm(n), n, 1)
Xt <- matrix(rnorm(nt*p, 0, 0.3), nt, p)
yt = matrix(rnorm(nt), nt, 1)
standardize(X, y, Xt, yt)


}
